% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_utils.R
\name{box_intercepts_dims}
\alias{box_intercepts_dims}
\title{Returns x and y coords of a bounding box}
\usage{
box_intercepts_dims(grid_data_plot, intersect_levels, box_pad = 0.375)
}
\arguments{
\item{grid_data_plot}{ggplot object, for example a plot created by \link[settleR]{grid_dot_plot} or similiar plots}

\item{intersect_levels}{Vector of intercepts which specify intercepts to draw a box around}

\item{box_pad}{padding surrounding the box}
}
\description{
Returns x and y coords of a bounding box
}
