% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_utils.R
\name{create_count_df}
\alias{create_count_df}
\title{Creates a data.frame with the counts per set}
\usage{
create_count_df(binary_mat)
}
\arguments{
\item{binary_mat}{binary matrix (see \link[settleR]{sets_to_matrix})}
}
\value{
a data.frame
}
\description{
Takes a binary matrix, for example the output from \link[settleR]{sets_to_matrix}
or similiar and returns a data.frame with the counts (or frequency) of the
set for each of the groups.
}
