% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleR_utils.R
\name{get_exlcusive_items}
\alias{get_exlcusive_items}
\title{Returns the items exclusive to a particular set}
\usage{
get_exlcusive_items(setList, setLevelCombo)
}
\arguments{
\item{setList}{list of sets}

\item{setLevelCombo}{combination of the setLevels (see \link[settleR]{SettleR})}
}
\value{
vector of items
}
\description{
Returns the items exclusive to a particular set
}
