% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{grid_dot_plot}
\alias{grid_dot_plot}
\title{Plotting showing the grids}
\usage{
grid_dot_plot(grid_data, dot_size = 6.25, colMap = NULL)
}
\arguments{
\item{grid_data}{data.frame with intersect_id, set_names, and observed as columns}

\item{dot_size}{size of dot}

\item{colMap}{optional named vector specifying text colours of y-axis labels}
}
\description{
Plotting showing the grids
}
