% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{intersect_bar_plot}
\alias{intersect_bar_plot}
\title{Creates bar plot of the intersects}
\usage{
intersect_bar_plot(
  intersect_data,
  plot_title = "Size of overlap",
  add_label = FALSE
)
}
\arguments{
\item{intersect_data}{data.frame with x, intersect_id, intersect_set_size, and intersect_degree as columns}
}
\value{
intersectPlotX
}
\description{
Creates bar plot of the intersects
}
\keyword{internal}
