% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleR-methods.R
\name{make_settleR_plots}
\alias{make_settleR_plots}
\title{Creates the 3 plots for a settleR plot}
\usage{
make_settleR_plots(settleRObject)
}
\arguments{
\item{settleRObject}{\link[settleR]{SettleR} object created by the constructor function.}
}
\description{
TODO incorperate this function into the SettleR validity checking
}
