% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{make_upset_plots}
\alias{make_upset_plots}
\title{Creates list of upset plots}
\usage{
make_upset_plots(upset_list, colMap = NULL)
}
\arguments{
\item{upset_list}{list of data.frames from \link[settleR]{calc_set_overlaps}}

\item{colMap}{optional named vector specifying text colours of y-axis labels in \link[settleR]{grid_dot_plot}}
}
\description{
This is a wrapper which just wraps the creation
of the 3 plots for required for an upset plot.
I've broken the plot creation into simple
functions to make further customizations
easy without re-writing everything
}
\keyword{internal}
