% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{merge_upset_list}
\alias{merge_upset_list}
\title{Uses patchwork to merge the plots}
\usage{
merge_upset_list(plt_list)
}
\arguments{
\item{plt_list}{}
}
\value{

}
\description{
This allows the marginal plots to have an axis. I couldn't do this before with
cowplot. I'm using this as default way to merge the plots together, but leaving
in the old funciton. The drawback of this approach is the inability to use
a fixed aspect ratio.
}
\keyword{internal}
