% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{set_panel_size}
\alias{set_panel_size}
\title{Fixes size of ggplot grob object}
\usage{
set_panel_size(g, width = 4, height = 4)
}
\arguments{
\item{g}{a gtable object returned from \code{\link[ggplot2]{ggplotGrob}}}

\item{width}{an int}

\item{height}{also an int}
}
\value{
gtable object
}
\description{
This sets the size of ggplot grob object to fixed
sizes. I've adapted this from \url{https://stackoverflow.com/a/30571289}.
The height and width parameters are in cm. May change in future.
}
\examples{

}
\keyword{internal}
