% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{set_totals_bar_plot}
\alias{set_totals_bar_plot}
\title{Creates ggplot bars of set size}
\usage{
set_totals_bar_plot(set_totals)
}
\arguments{
\item{set_totals}{}
}
\description{
bottom left plot which shows the size of the sets
}
