% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_utils.R
\name{sets_to_matrix}
\alias{sets_to_matrix}
\title{Converts list of sets into a binary matrix}
\usage{
sets_to_matrix(setList)
}
\arguments{
\item{setList}{list of sets}
}
\value{
binary matrix
}
\description{
This creates a matrix simliar to that in the UpSetR
package but I've removed a lot of the unneeded code.
I've also added a few helpful errors here.
}
