% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleRClass.R
\docType{class}
\name{SettleR-class}
\alias{SettleR-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{setList}}{list.}

\item{\code{binaryMat}}{matrix.}

\item{\code{countDf}}{data.frame.}

\item{\code{gridData}}{data.frame.}

\item{\code{intersectData}}{data.frame.}

\item{\code{setTotals}}{data.frame.}

\item{\code{plotList}}{list.}

\item{\code{setLevels}}{character.}

\item{\code{colMap}}{character.}

\item{\code{intersectLevels}}{character.}

\item{\code{nIntersects}}{numeric.}
}}

\examples{

}
