% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleRClass.R
\name{SettleR}
\alias{SettleR}
\title{Constructor for SettleR class}
\usage{
SettleR(
  setList,
  intersectLevels = NULL,
  setLevels = NULL,
  nIntersects = 15,
  colMap = NULL
)
}
\arguments{
\item{setList}{List of sets}

\item{intersectLevels}{Optional vector of intersects which specifies ordering of intersects. If null, intersects are sorted by size.}

\item{setLevels}{Optional vector of the set names (y-axis on \link[settleR]{grid_dot_plot}). If null, sets are ordered by size.}

\item{nIntersects}{Integer number of intersects to show (default: 15)}

\item{colMap}{Optional named vector specifying text colours of y-axis labels}
}
\value{
SettleR object
}
\description{
Constructor for SettleR class
}
