% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleR-methods.R
\name{settleR_plot}
\alias{settleR_plot}
\title{Plots the settleR plot on the device}
\usage{
settleR_plot(settleRObject)
}
\arguments{
\item{settleRObject}{}
}
\description{
Plots the settleR plot on the device
}
