% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{settleR_save}
\alias{settleR_save}
\title{wrapper to save the upset plot created by settleR_plot}
\usage{
settleR_save(upset_plt, fn)
}
\arguments{
\item{upset_plt}{an upset_plt gtabl object}

\item{fn}{file name}
}
\description{
Essentially, this is just a wrapper around ggsave. The difference is that
plot size is automatically calculated.
}
