% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{siteCount}
\alias{siteCount}
\title{Produce a time series that counts up or otherwise summarises the intensity of observed archaeological/historical sites or site phases per year}
\usage{
siteCount(startdates, enddates, siteprobs = NA, siteareas = NA,
  consensus = NA, conprop = 0.5, timeRange = NA, calendar = "BCAD",
  aoristic = FALSE, verbose = TRUE)
}
\arguments{
\item{startdates}{A vector of numeric values specifying the approximate start dates of each site or site phase.}

\item{enddates}{A vector of numeric values specifying the approximate end dates of each site or site phase.}

\item{siteprobs}{A vector of numeric values between 0 and 1 specifying the probability that the site in question was in use between the specified start and end date (often a quantification of how confidently an archaeological survey or excavation project were able to assign a given site to the specified chronological period based on recovered finds or other evidence).}

\item{siteareas}{A vector of numeric values >=0 that provide a size estimate of the sites or site phases (e.g. in hectares).}

\item{consensus}{A two column matrix or data.frame with a first column of years in the specified calendar, and a second column with an alternative set of settlement or population intensities (e.g. a summed radiocarbon distribution) for which a consensus model might be specified. This dataset must cover the full range of start and end dates, as well as of timeRange if this is given.}

\item{conprop}{Controls the importance of the consensus datset relative to the main sitecount dataset (default is 0.5 which imposes an equal weighting, but higher values would prioritise the consensus dataset more). Either a single number between 0 and 1 or a numeric vector of the same numebr of rows as consensus (each value between 0 and 1)}

\item{timeRange}{Numeric vector of length 2, specifying the earliest and latest date to include in the results (in the year units of the calendar argument). The default option will calculate this automatically from the input data.}

\item{calendar}{Specifies whether the input calendar is in years BC (-ve)/AD (+ve) or BP (default is "BCAD"). Note there should be no year 0 in "BCAD".}

\item{aoristic}{Logical value specifying whether an \"aoristic\" sum should be calculated.}

\item{verbose}{A logical variable indicating whether extra information on progress should be reported. Default is TRUE.}
}
\value{
A data.frame with the maximum number of sites or site phases counted for each year of the time range. Optionally, this dataframe may also have columns for a minimum count and middle-range, probability-weighted count, calculated based on the probability of site existence (siteprobs argument). Equivalent measures of maximum, minimum and probability-weighted total site area can also be calculated if the siteareas argument is specified. If aoristicsum is set to TRUE, then equivalent measures for an aoristic sum of sites can also be added. If a consensus dataset is provided, such as a radiocarbon summeed rpobability distribution, then an attempt is made (for site count, and if present, site area) to achieve a consensus between the settlement data and the consensus dataset. For example, for each site or site phases, it would assign the shape of the consensus dataset intensity for that timespan.
}
\description{
Given a set of archaeological/historical sites or site phases with absolute start and end dates, count up how many sites are definitely or possible present for each year in a specified time range.
}
\references{
#' Ratcliffe, J.H. 2000. Aoristic analysis: the spatial interpretation of unspecifed temporal events, International Journal of Geographical Information Science 14: 669–679.x#'
Johnson, I. 2004. Aoristic analysis: seeds of a new approach to mapping archaeological distributions through time, In K.F. Ausserer, W. Börner, M. Goriany and L. Karlhuber-Vöckl (eds.), [Enter the past]. The E-way into the Four Dimensions of Cultural Heritage (CAA 2003): 448–452. Oxford: Archaeopress.x#'
Palmisano, A., Bevan, A. and S. Shennan 2017. Comparing archaeological proxies for long-term population patterns: An example from central Italy, Journal of Archaeological Science 87: 59-72 448–452.x#'
}
