% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{upset_canvas_wrapper}
\alias{upset_canvas_wrapper}
\title{Wrapper for upset plot}
\usage{
upset_canvas_wrapper(upset_list, grid_size = 0.75)
}
\arguments{
\item{grid_size}{}
}
\description{
Wrapper for upset plot
}
