% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{upset_wrapper}
\alias{upset_wrapper}
\title{Wrapper to align and put the plots together}
\usage{
upset_wrapper(ol)
}
\arguments{
\item{ol}{}
}
\description{
Wrapper to align and put the plots together
}
