% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleRClass.R
\name{accessors}
\alias{accessors}
\alias{gridData}
\alias{gridData<-}
\alias{intersectData}
\alias{intersectData<-}
\alias{setTotals}
\alias{setTotals<-}
\alias{plotList}
\alias{plotList<-}
\alias{setLevels}
\alias{setLevels<-}
\alias{intersectLevels}
\alias{intersectLevels<-}
\alias{colMap}
\alias{colMap<-}
\alias{gridPlot}
\alias{gridPlot<-}
\alias{intersectPlotX}
\alias{intersectPlotX<-}
\alias{setPlotY}
\alias{setPlotY<-}
\title{Setters and Getters}
\usage{
gridData(object)

gridData(object) <- value

intersectData(object)

intersectData(object) <- value

setTotals(object)

setTotals(object) <- value

plotList(object)

plotList(object) <- value

setLevels(object)

setLevels(object) <- value

intersectLevels(object)

intersectLevels(object) <- value

colMap(object)

colMap(object) <- value

gridPlot(object)

gridPlot(object) <- value

intersectPlotX(object)

intersectPlotX(object) <- value

setPlotY(object)

setPlotY(object) <- value
}
\description{
Place to put all documentation for Setters and Getters
}
