% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{box_intercepts_dims}
\alias{box_intercepts_dims}
\title{Returns x and y coords of a bounding box for a particular intersect}
\usage{
box_intercepts_dims(grid_data_plot, intersects_to_box, box_pad = 0.5125)
}
\arguments{
\item{grid_data_plot}{ggplot object, for example a plot created by \link[settleR]{grid_dot_plot} or similiar plots}

\item{intersects_to_box}{Vector of intercepts which specify intercepts to draw a box around}

\item{box_pad}{padding surrounding the box}
}
\value{

}
\description{
Returns x and y coords of a bounding box for a particular intersect
}
