% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_utils.R
\name{calc_set_overlaps}
\alias{calc_set_overlaps}
\title{Creates df of upset data}
\usage{
calc_set_overlaps(binary_mat, intersectLevels = NULL, setLevels = NULL,
  nintersects = 15)
}
\arguments{
\item{binary_mat}{Binary matrix from \link[settleR]{sets_to_matrix}}

\item{intersectLevels}{Optional vector of intersects which specifies ordering of intersects. If null, intersects are sorted by size.}

\item{setLevels}{Optional vector of the set names (y-axis on \link[settleR]{grid_dot_plot}). If null, sets are ordered by size.}

\item{nintersects}{Integer number of intersects to show}
}
\description{
This formats the resulting data to long(ish)
formatted data. The reason why I've tried
put everything in 1 data.frame is so that
I can set the factors once in a single object.
It's currently adding more work to the uspset
wrapper since I have to essentially split this
into seperate objects, but I think it's worth
it so that I don't have to always think about the factors.
}
