% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_utils.R
\name{get_reordered_intersect_lvls}
\alias{get_reordered_intersect_lvls}
\title{Returns vector of intersect levels}
\usage{
get_reordered_intersect_lvls(setList, setLevels, nIntersects = 15)
}
\arguments{
\item{setList}{list of sets}

\item{setLevels}{Optional vector of the set names (y-axis on \link[settleR]{grid_dot_plot}). If null, sets are ordered by size.}

\item{nIntersects}{Integer number of intersects to show}
}
\description{
Intersect levels are reordered such that the
exclusive  singleton sets are the first n
number of intersects and the remaining are
ordered from largest to smallest like normal.
}
