% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{grid_dot_plot}
\alias{grid_dot_plot}
\title{Plotting showing the grids}
\usage{
grid_dot_plot(
  grid_data,
  dot_size = 4.25,
  plot_title = "Overlap index",
  colMap = NULL
)
}
\arguments{
\item{grid_data}{data.frame with intersect_id, set_names, and observed as columns}

\item{dot_size}{size of dot}

\item{colMap}{optional named vector specifying text colours of y-axis labels}
}
\value{
gridPlot
}
\description{
Plotting showing the grids
}
\keyword{internal}
