% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{grid_dot_plot}
\alias{grid_dot_plot}
\title{Plotting showing the grids}
\usage{
grid_dot_plot(grid_data, dot_size = 5)
}
\arguments{
\item{grid_data}{}
}
\description{
Plotting showing the grids
}
