% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{intersect_bar_plot}
\alias{intersect_bar_plot}
\title{Creates bar plot of the intersects}
\usage{
intersect_bar_plot(intersect_data)
}
\arguments{
\item{intersect_data}{}
}
\description{
Creates bar plot of the intersects
}
