% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_plotting_functions.R
\name{make_settleR_plots}
\alias{make_settleR_plots}
\title{Creates the 3 plots for a settleR plot}
\usage{
make_settleR_plots(settleRObject)
}
\arguments{
\item{settleRObject}{}
}
\description{
Creates the 3 plots for a settleR plot
}
