% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{merge_upset_list}
\alias{merge_upset_list}
\title{Merges list of ggplot upset into gtable object}
\usage{
merge_upset_list(upset_list, grid_size = 0.75)
}
\arguments{
\item{upset_list}{list of ggplots created from \link[settleR]{make_upset_plots} or similiarly structured}

\item{grid_size}{sets grid size}
}
\description{
Merges list of ggplot upset into gtable object
}
