% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SettleR-methods.R
\name{reorder_by_singletons}
\alias{reorder_by_singletons}
\title{Returns vector of intersect levels}
\usage{
reorder_by_singletons(settleRObject)
}
\arguments{
\item{settleRObject}{\link[settleR]{SettleR} object created by the constructor function.}
}
\value{

}
\description{
Intersect levels are reordered such that the
exclusive  singleton sets are the first n
number of intersects and the remaining are
ordered from largest to smallest like normal.
}
