% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{set_totals_bar_plot}
\alias{set_totals_bar_plot}
\title{Creates ggplot bars of set size}
\usage{
set_totals_bar_plot(set_totals)
}
\arguments{
\item{set_totals}{data.frame with y, set_names, total_set_size as columns}
}
\value{
setPlotY
}
\description{
bottom left plot which shows the size of the sets
}
\keyword{internal}
