% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settleR_utils.R
\name{sets_to_matrix}
\alias{sets_to_matrix}
\title{Converts list of sets into a binary matrix}
\usage{
sets_to_matrix(setlist)
}
\arguments{
\item{setlist}{list of sets}
}
\value{
binary matrix
}
\description{
Converts list of sets into a binary matrix
}
