% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{settleR_plot}
\alias{settleR_plot}
\title{Creates a settleR plot starting from a named list of sets}
\usage{
settleR_plot(setlist, nintersects = 15, set_levels = NULL,
  col_map = NULL, grid_size = 0.75)
}
\arguments{
\item{setlist}{list of sets}

\item{nintersects}{Integer number of intersects to show}

\item{set_levels}{Optional vector of the set names (y-axis on \link[settleR]{grid_dot_plot}). If null, sets are ordered by size.}

\item{col_map}{optional named vector specifying text colours of y-axis labels}

\item{grid_size}{sets grid size}
}
\description{
Creates a settleR plot starting from a named list of sets
}
