% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{settleR_plot}
\alias{settleR_plot}
\title{Creates a settleR plot starting from a named list of sets}
\usage{
settleR_plot(setList, nIntersects = 15, setLevels = NULL,
  colMap = NULL, margSize = 5)
}
\arguments{
\item{setList}{list of sets}

\item{nIntersects}{Integer number of intersects to show}

\item{setLevels}{Optional vector of the set names (y-axis on \link[settleR]{grid_dot_plot}). If null, sets are ordered by size.}

\item{colMap}{optional named vector specifying text colours of y-axis labels}

\item{margSize}{sets height or width of plots in margin size in cm}
}
\description{
Creates a settleR plot starting from a named list of sets
}
