% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{theme_empty}
\alias{theme_empty}
\title{Modified ggplot2 theme}
\usage{
theme_empty(base_size = 15)
}
\arguments{
\item{base_size}{font size}
}
\value{

}
\description{
This just returns a modifed version of \link[ggplot2]{theme_void}
}
